Move some methods and fields around
