+++ cardinal-components-chunk/src/main/java/nerdhub/cardinal/components/internal/StaticChunkComponentPlugin.java
@@ -25,7 +25,7 @@ package nerdhub.cardinal.components.internal;
 import nerdhub.cardinal.components.api.ChunkComponentFactory;
 import net.fabricmc.loader.api.FabricLoader;
 
-public final class StaticChunkComponentPlugin extends SimpleStaticComponentPlugin {
+public final class StaticChunkComponentPlugin extends StaticComponentPluginBase {
     public static final String CHUNK_IMPL_SUFFIX = "ChunkImpl";
     public static final String CHUNK_CLASS = FabricLoader.getInstance().getMappingResolver().mapClassName("intermediary", "net.minecraft.class_2791");
